rest module contains only REST model / interfaces
only classes / interfaces needed to interact with bridge left, implementation moved to core
