+++ core/src/main/java/org/jboss/pnc/causeway/brewclient/BrewClient.java
@@ -1,8 +1,8 @@
 package org.jboss.pnc.causeway.brewclient;
 
-import org.jboss.pnc.causeway.model.BrewBuild;
-import org.jboss.pnc.causeway.model.BrewNVR;
-import org.jboss.pnc.causeway.model.BuildImportResult;
+import org.jboss.pnc.causeway.rest.BrewBuild;
+import org.jboss.pnc.causeway.rest.BrewNVR;
+import org.jboss.pnc.causeway.rest.BuildImportResult;
 import org.jboss.pnc.causeway.pncclient.PncBuild;
 
 import javax.enterprise.context.ApplicationScoped;
