+++ core/src/main/java/org/jboss/pnc/causeway/ctl/PncImportController.java
@@ -2,10 +2,10 @@ package org.jboss.pnc.causeway.ctl;
 
 import org.jboss.pnc.causeway.CausewayException;
 import org.jboss.pnc.causeway.brewclient.BrewClient;
-import org.jboss.pnc.causeway.model.BuildImportResult;
-import org.jboss.pnc.causeway.model.BrewBuild;
-import org.jboss.pnc.causeway.model.BrewNVR;
-import org.jboss.pnc.causeway.model.ProductReleaseImportResult;
+import org.jboss.pnc.causeway.rest.BuildImportResult;
+import org.jboss.pnc.causeway.rest.BrewBuild;
+import org.jboss.pnc.causeway.rest.BrewNVR;
+import org.jboss.pnc.causeway.rest.ProductReleaseImportResult;
 import org.jboss.pnc.causeway.pncclient.PncBuild;
 import org.jboss.pnc.causeway.pncclient.PncClient;
