+++ core/src/main/java/org/jboss/pnc/causeway/rest/PncImportResourceEndpoint.java
@@ -1,0 +1,41 @@
+package org.jboss.pnc.causeway.rest;
+
+import org.commonjava.propulsor.deploy.resteasy.RestResources;
+import org.jboss.pnc.causeway.CausewayException;
+import org.jboss.pnc.causeway.ctl.PncImportController;
+import org.jboss.pnc.causeway.rest.ProductReleaseImportResult;
+
+import javax.enterprise.context.RequestScoped;
+import javax.inject.Inject;
+import javax.ws.rs.GET;
+import javax.ws.rs.Path;
+import javax.ws.rs.PathParam;
+import javax.ws.rs.WebApplicationException;
+import javax.ws.rs.core.Response;
+
+@RequestScoped
+public class PncImportResourceEndpoint
+        implements RestResources, PncImportResource {
+
+    @Inject
+    private PncImportController controller;
+
+    @GET
+    @Path( "test/{variable}" )
+    public Response testResponse( @PathParam( "variable" ) String var )
+    {
+        return Response.ok( var ).build();
+    }
+
+    public ProductReleaseImportResult importProductRelease(int releaseId )
+    {
+        try
+        {
+            return controller.importProductRelease( releaseId );
+        }
+        catch ( CausewayException e )
+        {
+            throw new WebApplicationException( e, Response.Status.INTERNAL_SERVER_ERROR );
+        }
+    }
+}
