+++ rest/pom.xml
@@ -29,36 +29,20 @@
   
   <dependencies>
     <dependency>
-      <groupId>org.jboss.pnc.causeway</groupId>
-      <artifactId>causeway-model-java</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.jboss.pnc.causeway</groupId>
-      <artifactId>causeway-core</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.commonjava.propulsor</groupId>
-      <artifactId>propulsor-resteasy</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.jboss.pnc</groupId>
-      <artifactId>rest-model</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>commons-lang</groupId>
-      <artifactId>commons-lang</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>commons-io</groupId>
-      <artifactId>commons-io</artifactId>
-    </dependency>
-    <dependency>
       <groupId>com.fasterxml.jackson.core</groupId>
       <artifactId>jackson-core</artifactId>
     </dependency>
     <dependency>
       <groupId>com.fasterxml.jackson.core</groupId>
       <artifactId>jackson-databind</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.jboss.resteasy</groupId>
+      <artifactId>jaxrs-api</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>com.fasterxml.jackson.core</groupId>
+      <artifactId>jackson-annotations</artifactId>
     </dependency>
     <dependency>
       <groupId>junit</groupId>
