+++ rest/src/main/java/org/jboss/pnc/causeway/rest/BrewNVR.java
@@ -1,4 +1,4 @@
-package org.jboss.pnc.causeway.model;
+package org.jboss.pnc.causeway.rest;
 
 public class BrewNVR {
