+++ rest/src/test/java/org/jboss/pnc/causeway/ImportedBuildJsonConversionIntegrationTest.java
@@ -2,9 +2,9 @@ package org.jboss.pnc.causeway;
 
 import static org.junit.Assert.assertEquals;
 
-import org.jboss.pnc.causeway.model.BrewBuild;
-import org.jboss.pnc.causeway.model.BrewNVR;
-import org.jboss.pnc.causeway.model.ImportedBuild;
+import org.jboss.pnc.causeway.rest.BrewBuild;
+import org.jboss.pnc.causeway.rest.BrewNVR;
+import org.jboss.pnc.causeway.rest.ImportedBuild;
 import org.junit.Test;
 
 import java.io.ByteArrayOutputStream;
